% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MaxPool2d}
\alias{MaxPool2d}
\title{Max Pooling function}
\usage{
MaxPool2d(hidden)
}
\arguments{
\item{hidden}{the matrix of the neurons after the activation function}
}
\value{
the matrix after Pooling,the size is half of the previous
}
\description{
This function is to reduce dimensionality and prevent overfitting
}
\examples{
\dontrun{
MaxPool2d(hidden = matrix(data = rnorm(100), ncol = 10))
}
}
