% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simple_conv2d}
\alias{simple_conv2d}
\title{Convolution function}
\usage{
simple_conv2d(img, kernel_size)
}
\arguments{
\item{img}{the matrix of picture pixel}

\item{kernel_size}{is given according to the experimental requirements,and kernel, the filter is generated randomly}
}
\value{
the matrix after convolution,the size is ((img_nrow - kernel_row)/stride + 1)*((img_col - kernel_col)/stride + 1)
}
\description{
Simple convolution function is to extract features with using filters(kernel),here only consider the case where the number of channels is 1, that is, to process gray images
}
\examples{
\dontrun{
simple_conv2d(img = matrix(data = rnorm(100), ncol = 10), kernel_size = c(4, 4))
}
}
